refactor(api): agents.py 분할 완료 — core·ocr 추출, _legacy 소멸
1,731줄 단일 파일을 6개 모듈로 분할 완료(orchestrate 476 · tmux 412 ·
ocr 404 · mcp 322 · core 200 · _shared 17 · __init__ 51).
HTTP 표면 무변경(라우트 29건 characterization 통과, 가림 쌍 0건),
소비자 import 경로 무변경.

_legacy.py 를 core.py 로 rename 했다. 남은 5개 라우트(list·stats·get_agent·
search·update_status)가 전부 core 도메인이다.

**집계 라우터를 core 모듈이 소유한다** — projects/registry 와 같은 이유다.
@router.get("")(list_agents) 처럼 경로가 빈 라우트가 있어 prefix 까지 비면
FastAPI 가 거부한다. 경로를 "/" 로 바꾸는 것은 동작 변경이므로 하지 않는다.

문자열 패치 타깃 6곳을 api.agents._legacy.X → api.agents.ocr.X 로 갱신했다.
여섯 개 전부 OCR 경로(_record_ocr_usage·OCR 핸들러·비전 모델 헬퍼)에 있어
ocr 모듈 하나로 수렴했다. 이로써 _legacy 를 가리키는 패치 지점은 0건이다.

게이트: 라우트 29건 동일, mypy 0, pytest 1348 passed(플레이크 1건 제외),
ruff·format 0.
