+++ src/backend/api/agents/ocr.py
@@ -1,4 +1,11 @@
-"""Agent API routes - Agent Registry, Lead Orchestrator, MCP Manager."""
+"""ocr 관련 Agents API 라우트.
+
+Task Analyzer 용 이미지 텍스트 추출(OCR). 비전 모델 런타임 해석과
+LLM 사용량 기록 헬퍼가 함께 산다 — 전부 이 핸들러 전용이다.
+
+`__init__.py` 가 `_resolve_ocr_runtime` 과 `extract_text_from_image` 를
+재노출한다(패키지 밖 테스트가 각각 import·모듈 속성으로 가져간다).
+"""
 
 import os
 import time
@@ -6,7 +13,7 @@ from pathlib import Path
 from typing import Any
 
 from fastapi import APIRouter, Depends, File, HTTPException, UploadFile
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from api.deps import get_current_user, get_db_session
@@ -20,12 +27,6 @@ from models.llm_usage import (
     LLMUsageRecordCreate,
     LLMUsageSource,
     LLMUsageStatus,
-)
-from services.agent_registry import (
-    AgentCategory,
-    AgentMetadata,
-    AgentStatus,
-    get_agent_registry,
 )
 from services.llm_access_service import get_access_for_user
 from services.llm_runtime_resolver import (
@@ -47,52 +48,7 @@ from ._shared import ALLOWED_IMAGE_EXTENSIONS, MAX_IMAGE_SIZE
 router = APIRouter(prefix="/agents", tags=["agents"])
 
 
-# ─────────────────────────────────────────────────────────────
-# Response Models
-# ─────────────────────────────────────────────────────────────
-
-
-class AgentResponse(BaseModel):
-    """에이전트 정보 응답."""
-
-    id: str
-    name: str
-    description: str
-    category: str
-    status: str
-    capabilities: list[dict[str, Any]]
-    specializations: list[str]
-    estimated_cost_per_task: float
-    avg_execution_time_ms: int
-    total_tasks_completed: int
-    success_rate: float
-    is_available: bool
-
-
-class AgentRegistryStatsResponse(BaseModel):
-    """에이전트 레지스트리 통계 응답."""
-
-    total_agents: int
-    available_agents: int
-    busy_agents: int
-    by_category: dict[str, int]
-    total_tasks_completed: int
-    avg_success_rate: float
-
-
-class AgentSearchRequest(BaseModel):
-    """에이전트 검색 요청."""
-
-    query: str = Field(..., description="검색 쿼리 (태스크 설명)")
-    category: str | None = Field(None, description="카테고리 필터")
-    limit: int = Field(5, description="최대 결과 수")
-
-
-class AgentSearchResult(BaseModel):
-    """에이전트 검색 결과."""
-
-    agent: AgentResponse
-    score: int
+router = APIRouter()
 
 
 class OCRResponse(BaseModel):
@@ -285,77 +241,6 @@ async def _record_ocr_usage(
     )
 
 
-# ─────────────────────────────────────────────────────────────
-# Agent Registry API
-# ─────────────────────────────────────────────────────────────
-
-
-def _agent_to_response(agent: AgentMetadata) -> AgentResponse:
-    """AgentMetadata를 AgentResponse로 변환."""
-    return AgentResponse(
-        id=agent.id,
-        name=agent.name,
-        description=agent.description,
-        category=agent.category.value,
-        status=agent.status.value,
-        capabilities=[
-            {
-                "name": cap.name,
-                "description": cap.description,
-                "keywords": cap.keywords,
-                "priority": cap.priority,
-            }
-            for cap in agent.capabilities
-        ],
-        specializations=agent.specializations,
-        estimated_cost_per_task=agent.estimated_cost_per_task,
-        avg_execution_time_ms=agent.avg_execution_time_ms,
-        total_tasks_completed=agent.total_tasks_completed,
-        success_rate=agent.success_rate,
-        is_available=agent.is_available(),
-    )
-
-
-@router.get("", response_model=list[AgentResponse])
-async def list_agents(
-    category: str | None = None,
-    available_only: bool = False,
-    _user: UserModel = Depends(get_current_user),
-):
-    """
-    모든 에이전트 목록 조회.
-
-    Args:
-        category: 카테고리 필터 (development, orchestration, quality, research)
-        available_only: 사용 가능한 에이전트만 반환
-    """
-    registry = get_agent_registry()
-
-    if category:
-        try:
-            cat = AgentCategory(category)
-            agents = registry.get_by_category(cat)
-        except ValueError:
-            raise HTTPException(
-                status_code=400,
-                detail=f"Invalid category: {category}. Valid: {[c.value for c in AgentCategory]}",
-            )
-    elif available_only:
-        agents = registry.get_available()
-    else:
-        agents = registry.get_all()
-
-    return [_agent_to_response(a) for a in agents]
-
-
-@router.get("/stats", response_model=AgentRegistryStatsResponse)
-async def get_registry_stats(_user: UserModel = Depends(get_current_user)):
-    """에이전트 레지스트리 통계 조회."""
-    registry = get_agent_registry()
-    stats = registry.get_stats()
-    return AgentRegistryStatsResponse(**stats)
-
-
 @router.post("/ocr", response_model=OCRResponse)
 async def extract_text_from_image(
     image: UploadFile = File(..., description="OCR 대상 이미지"),
@@ -517,67 +402,3 @@ async def extract_text_from_image(
             error=f"텍스트 추출 실패: {str(e)}",
             model_used=runtime_resolution.model_id,
         )
-
-
-@router.get("/{agent_id}", response_model=AgentResponse)
-async def get_agent(agent_id: str, _user: UserModel = Depends(get_current_user)):
-    """특정 에이전트 조회."""
-    registry = get_agent_registry()
-    agent = registry.get(agent_id)
-
-    if not agent:
-        raise HTTPException(status_code=404, detail=f"Agent not found: {agent_id}")
-
-    return _agent_to_response(agent)
-
-
-@router.post("/search", response_model=list[AgentSearchResult])
-async def search_agents(request: AgentSearchRequest, _user: UserModel = Depends(get_current_user)):
-    """
-    능력 기반 에이전트 검색.
-
-    태스크 설명을 기반으로 가장 적합한 에이전트를 찾습니다.
-    """
-    registry = get_agent_registry()
-
-    category = None
-    if request.category:
-        try:
-            category = AgentCategory(request.category)
-        except ValueError:
-            raise HTTPException(status_code=400, detail=f"Invalid category: {request.category}")
-
-    results = registry.find_by_capability(
-        query=request.query,
-        category=category,
-        limit=request.limit,
-    )
-
-    return [
-        AgentSearchResult(
-            agent=_agent_to_response(agent),
-            score=score,
-        )
-        for agent, score in results
-    ]
-
-
-@router.post("/{agent_id}/status")
-async def update_agent_status(
-    agent_id: str, status: str, _user: UserModel = Depends(get_current_user)
-):
-    """에이전트 상태 업데이트."""
-    registry = get_agent_registry()
-
-    try:
-        new_status = AgentStatus(status)
-    except ValueError:
-        raise HTTPException(
-            status_code=400,
-            detail=f"Invalid status: {status}. Valid: {[s.value for s in AgentStatus]}",
-        )
-
-    if not registry.update_status(agent_id, new_status):
-        raise HTTPException(status_code=404, detail=f"Agent not found: {agent_id}")
-
-    return {"message": f"Agent {agent_id} status updated to {status}"}
