+++ tests/backend/test_llm_usage_instrumentation.py
@@ -509,13 +509,13 @@ async def test_task_analyzer_ocr_records_allowed_api_fallback_usage(monkeypatch)
     recorder = AsyncMock()
     preflight = AsyncMock()
 
-    monkeypatch.setattr("api.agents._legacy.get_access_for_user", AsyncMock(return_value=access))
-    monkeypatch.setattr("api.agents._legacy.LLMModelRegistry.get_enabled", MagicMock(return_value=[model]))
-    monkeypatch.setattr("api.agents._legacy.LLMModelRegistry.is_available", MagicMock(return_value=True))
-    monkeypatch.setattr("api.agents._legacy.LLMService._get_llm", MagicMock(return_value=fake_llm))
-    monkeypatch.setattr("api.agents._legacy.record_usage_best_effort", recorder)
+    monkeypatch.setattr("api.agents.ocr.get_access_for_user", AsyncMock(return_value=access))
+    monkeypatch.setattr("api.agents.ocr.LLMModelRegistry.get_enabled", MagicMock(return_value=[model]))
+    monkeypatch.setattr("api.agents.ocr.LLMModelRegistry.is_available", MagicMock(return_value=True))
+    monkeypatch.setattr("api.agents.ocr.LLMService._get_llm", MagicMock(return_value=fake_llm))
+    monkeypatch.setattr("api.agents.ocr.record_usage_best_effort", recorder)
     monkeypatch.setattr(
-        "api.agents._legacy.enforce_usage_quota_preflight_best_effort",
+        "api.agents.ocr.enforce_usage_quota_preflight_best_effort",
         preflight,
     )
